test(lsp): move folding range tests
